5223: Create a resource to submit AuthnRequests to the Connector Node
For now we're just passing the ID of the incoming SAML request in the
URL, because this needs to be the same in the message we send to the
connector node. Later this will want to go in the user's session (at the
moment the plan is for this to be a JWT).

Added AppRule test for the new resource and unit tests for the
EidasAuthnRequestGenerator.

TODO: all the rest of the AuthnRequest.

Authors: @richardtowers @suneetkamath