5223: Build our own RequestedAttribute implementation
Eidas add some SAML extensions which need to be marshalled to and from
XML. There's a RequestedAttribute element which is similar to a
Attribute from opensaml core, but adds an isRequired attribute. This
commit adds the implementation of the opensaml classes needed to
represent, marshal, and unmarshal these elements.

We've also put the eidas XML Namespace on the root AuthnRequest element
Rather than declaring it where it's first used. This is neater, and it
makes the AuthnRequest slightly smaller because we don't have to declare
the namespace multiple times.

Authors: @richardtowers @sfkamath