5223: Sign Eidas AuthnRequests, set issuer to metadata URI
The keystore containing the signing key will be provided as a base64
encoded byte array in the configuration.

We've also changed the issuer to be the URL of our metadata, because it
seems that this is the thing eIDAS uses to fetch metadata and thus check
the signature on the AuthnRequest.

To check the signature in the integration tests we need to dynamically
build a keystore (see TestSigningKeyStoreProvider), which is quite
complex.

Authors: @richardtowers @sfkamath @chrisholmes